chore: 提交所有最新修改，包括合约、脚本、测试、文档等全量同步

+++ script/Deploy.s.sol
@@ -8,11 +8,7 @@ contract DeployScript is Script {
     DCAllocator public dcAllocator;
 
     // 从环境变量中读取配置
-    function getConfig() internal view returns (
-        uint256 threshold,
-        address vault,
-        uint256 challengePeriod
-    ) {
+    function getConfig() internal view returns (uint256 threshold, address vault, uint256 challengePeriod) {
         // 从环境变量获取阈值
         threshold = vm.envOr("THRESHOLD", uint256(2));
         require(threshold > 0, "Invalid threshold");
@@ -29,45 +25,37 @@ contract DeployScript is Script {
     }
 
     function run() public {
-        (
-            uint256 threshold,
-            address vault,
-            uint256 challengePeriod
-        ) = getConfig();
-        
+        (uint256 threshold, address vault, uint256 challengePeriod) = getConfig();
+
         console.log("Deploying DCAllocator with the following configuration:");
         console.log("Threshold:", threshold);
         console.log("Vault:", vault);
         console.log("Challenge Period (days):", challengePeriod);
-        
+
         vm.startBroadcast();
-        
+
         address multisig = 0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc; // 示例多签地址
-        dcAllocator = new DCAllocator(
-            vault,
-            challengePeriod,
-            multisig
-        );
-        
+        dcAllocator = new DCAllocator(vault, challengePeriod, multisig);
+
         console.log("DCAllocator deployed at:", address(dcAllocator));
-        
+
         vm.stopBroadcast();
     }
-    
+
     // 辅助函数：将字符串按分隔符分割
     function splitString(string memory str, string memory delimiter) internal pure returns (string[] memory) {
         bytes memory strBytes = bytes(str);
         bytes memory delimiterBytes = bytes(delimiter);
-        
+
         if (strBytes.length == 0) {
             string[] memory empty = new string[](0);
             return empty;
         }
-        
-        uint count = 1;
-        for (uint i = 0; i < strBytes.length; i++) {
+
+        uint256 count = 1;
+        for (uint256 i = 0; i < strBytes.length; i++) {
             bool isDelimiter = true;
-            for (uint j = 0; j < delimiterBytes.length; j++) {
+            for (uint256 j = 0; j < delimiterBytes.length; j++) {
                 if (i + j >= strBytes.length || strBytes[i + j] != delimiterBytes[j]) {
                     isDelimiter = false;
                     break;
@@ -78,54 +66,57 @@ contract DeployScript is Script {
                 i += delimiterBytes.length - 1;
             }
         }
-        
+
         string[] memory parts = new string[](count);
-        uint partIndex = 0;
-        uint startIndex = 0;
-        
-        for (uint i = 0; i <= strBytes.length; i++) {
-            if (i == strBytes.length || (i <= strBytes.length - delimiterBytes.length && isSubstringAt(strBytes, delimiterBytes, i))) {
-                uint length = i - startIndex;
+        uint256 partIndex = 0;
+        uint256 startIndex = 0;
+
+        for (uint256 i = 0; i <= strBytes.length; i++) {
+            if (
+                i == strBytes.length
+                    || (i <= strBytes.length - delimiterBytes.length && isSubstringAt(strBytes, delimiterBytes, i))
+            ) {
+                uint256 length = i - startIndex;
                 bytes memory part = new bytes(length);
-                for (uint j = 0; j < length; j++) {
+                for (uint256 j = 0; j < length; j++) {
                     part[j] = strBytes[startIndex + j];
                 }
                 parts[partIndex] = string(part);
                 partIndex++;
-                
+
                 if (i < strBytes.length) {
                     i += delimiterBytes.length - 1;
                     startIndex = i + 1;
                 }
             }
         }
-        
+
         return parts;
     }
-    
+
     // 辅助函数：检查子字符串是否在指定位置
-    function isSubstringAt(bytes memory str, bytes memory substr, uint pos) internal pure returns (bool) {
+    function isSubstringAt(bytes memory str, bytes memory substr, uint256 pos) internal pure returns (bool) {
         if (pos + substr.length > str.length) {
             return false;
         }
-        
-        for (uint i = 0; i < substr.length; i++) {
+
+        for (uint256 i = 0; i < substr.length; i++) {
             if (str[pos + i] != substr[i]) {
                 return false;
             }
         }
-        
+
         return true;
     }
-    
+
     // 辅助函数：将字符串解析为地址
     function parseAddr(string memory s) internal pure returns (address) {
         bytes memory ss = bytes(s);
         require(ss.length == 42, "Invalid address length");
-        require(ss[0] == '0' && ss[1] == 'x', "Invalid address format");
-        
+        require(ss[0] == "0" && ss[1] == "x", "Invalid address format");
+
         uint160 addr = 0;
-        for (uint i = 2; i < 42; i++) {
+        for (uint256 i = 2; i < 42; i++) {
             addr *= 16;
             uint8 digit = uint8(ss[i]);
             if (digit >= 48 && digit <= 57) {
@@ -141,7 +132,7 @@ contract DeployScript is Script {
                 revert("Invalid address character");
             }
         }
-        
+
         return address(addr);
     }
 }
